% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{breakFile}
\alias{breakFile}
\title{breakFile}
\usage{
breakFile(input, output, n, filename, method)
}
\arguments{
\item{input}{the file path/name of the csv you wish to break.}

\item{output}{the folder where your broken files will be written to.}

\item{n}{the number of rows you want your file to contain if your method is "row" or the number of files you want your original file broken into if your method is "fraction".}

\item{filename}{the name of your output files. Iterative numbers will be appended to this filename.}

\item{method}{can either be "row" or "fraction".}
}
\value{
returns files written to specified directory.
}
\description{
breakFile
}
\examples{
\dontrun{
# SET WORKING DIRECTORY
wd <- "C:/YOURFILEPATH/"

# SET INPUT FILE
dataSource <- paste(wd, "Input/Budget.csv", sep='')

# SET OUTPUT FOLDER
outwd <- paste(wd, "Output/Budget/", sep='')

breakFile(dataSource, outwd, 1000, "Budget", "row")}
}
