% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{breakRow}
\alias{breakRow}
\title{breakRow}
\usage{
breakRow(input, output, n, filename)
}
\arguments{
\item{input}{the file path/name of the csv you wish to break.}

\item{output}{the folder where your broken files will be written to.}

\item{n}{the number of rows you want your broken files to contain.}

\item{filename}{the name of your output files. Iterative numbers will be appended to this filename.}
}
\value{
returns files written to specified directory.
}
\description{
breakRow
}
\examples{
\dontrun{
#SET WORKING DIRECTORY
wd <- "C:/YOURFILEPATH/"

#SET INPUT FILE
dataSource <- paste(wd, "Input/Budget.csv", sep='')

#SET OUTPUT FOLDER
outwd <- paste(wd, "Output/Budget/", sep='')

breakRow(dataSource, outwd, 1000, "Budget")}
}
