% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{combineFiles}
\alias{combineFiles}
\title{combineFiles}
\usage{
combineFiles(input)
}
\arguments{
\item{input}{the working directory where your files are located.}
}
\value{
returns a dataframe named 'df' containing all of your combined files.
}
\description{
combineFiles
}
\examples{
\dontrun{
wd <- "C:/YOURWORKINGDIRECTORY"
combineFiles(wd)}
}
